sign_validation/requirements.txt
python==3.7
absl-py==0.11.0
anaconda-client==1.7.2
anaconda-navigator==1.9.2
appdirs==1.4.3
asgiref==3.2.10
attrs==18.2.0
Automat==0.7.0
backcall==0.1.0
beautifulsoup4==4.6.3
bleach==2.1.4
cachetools==4.1.1
certifi==2020.12.5
cffi==1.11.5
chardet==3.0.4
clyent==1.2.2
colorama==0.3.9
constantly==15.1.0
cycler==0.10.0
decorator==4.3.0
defusedxml==0.5.0
Django==3.1.2
django-crispy-forms==1.10.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.13
entrypoints==0.2.3
filelock==3.0.8
fsspec==0.8.4
future==0.18.2
glob2==0.6
grpcio==1.33.2
html5lib==1.0.1
hyperlink==18.0.0
idna==2.7
importlib-metadata==3.1.0
incremental==17.5.0
ipykernel==4.10.0
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.1
jedi==0.12.1
Jinja2==2.10
jsonschema==2.6.0
mkl-fft==1.0.4
mkl-random==1.0.1
navigator-updater==0.2.1
nbconvert==5.4.0
nbformat==4.4.0
notebook==5.6.0
numpy==1.19.4
oauthlib==3.1.0
olefile==0.46
opencv-python==4.4.0.46
pandocfilters==1.4.2
parso==0.3.1
pickleshare==0.7.4
Pillow==8.0.1
pkginfo==1.4.2
prometheus-client==0.3.1
prompt-toolkit==1.0.15
protobuf==3.14.0
psutil==5.4.7
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycosat==0.6.3
pycparser==2.18
Pygments==2.2.0
pyOpenSSL==18.0.0
PySocks==1.6.8
python-dateutil==2.7.3
pywin32==223
pywinpty==0.5.4
PyYAML==3.13
pyzmq==17.1.2
QtPy==1.5.0
requests-oauthlib==1.3.0
rsa==4.6
ruamel-yaml==0.15.46
scikit-learn==0.19.2
scipy==1.1.0
Send2Trash==1.5.0
service-identity==17.0.0
simplegeneric==0.8.1
sqlparse==0.4.1
TBB==0.1
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
terminado==0.8.1
testpath==0.3.1
torch ==1.8.0
torchvision ==0.9.0
tornado==5.1
traitlets==4.3.2
Twisted==18.7.0
urllib3==1.23
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.1
win-inet-pton==1.0.1
wincertstore==0.2
z3-solver==4.8.10.0
zipp==3.4.0
zope.interface==4.5.0
